+++ src/main/resources/db/migration/V2026.03.17.00.00__create_fcm_token_topic.sql
@@ -3,8 +3,8 @@ CREATE TABLE fcm_token_topic
     id           BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
     fcm_token_id BIGINT       NOT NULL,
     topic_name   VARCHAR(255) NOT NULL,
-    created_at   TIMESTAMP    NOT NULL DEFAULT NOW(),
-    updated_at   TIMESTAMP    NOT NULL DEFAULT NOW(),
+    created_at   TIMESTAMP(6) WITH TIME ZONE NOT NULL,
+    updated_at   TIMESTAMP(6) WITH TIME ZONE NOT NULL,
     CONSTRAINT fk_fcm_token_topic_fcm_token FOREIGN KEY (fcm_token_id) REFERENCES fcm_token (id) ON DELETE CASCADE
 );
